# src/main/webapp/docs/swagger.yaml
swagger: '2.0'
info:
  description: Api para buscar contenido audiovisual
  version: 1.0.0
  title: MovieFinder
  # put the contact info for your development or API team
  contact:
    email: alvagualh@alum.us.es
servers:
  url: https://snmanager.appspot.com/api
  description: Entrada a la api desplegada
# tags are used for organizing operations
...
paths:
...
  /songs:
    get:
      tags:
      - Song
      summary: getSongs
      operationId: getSongs
      description: Devuelve todas las canciones
      produces:
      - application/json
      responses:
        200:
          description: Dando todas las canciones
          schema:
            $ref: '#/definitions/Song'
    post:
      tags:
      - Song
      summary: addSong
      operationId: addSong
      description: Añadir una nueva cancion
      parameters:
      - in: body
        name: addSong
        description: Song to add
        schema:
          $ref: '#/definitions/Song'
      produces:
      - application/json
      responses:
        201:
          description: Created
    put:
      tags:
      - Song
      summary: updateSong
      operationId: updateSong
      description: Actualizar una cancion
      parameters:
      - in: body
        name: updateSong
        description: Song to add
        schema:
          $ref: '#/definitions/Song'
      produces:
      - application/json
      responses:
        204:
          description: No contend
        404:
          description: Not found
  /songs/{songId}:
    get:
      tags:
      - Song
      summary: getSong
      parameters:
      - in: path
        name: songId
        required: true
        schema:
          type: string
        description: Id de la cancion
      operationId: getSong
      description: Devuelve la cancion de ese id
      produces:
      - application/json
      responses:
        200:
          description: Dando resultados
          schema:
            $ref: '#/definitions/Song'
        404:
          description: Not found
    delete:
      tags:
      - Song
      summary: deleteSong
      parameters:
      - in: path
        name: songId
        required: true
        schema:
          type: string
        description: Id de la cancion
      operationId: deleteSong
      description: Elimina la cancion con ese id
      produces:
      - application/json
      responses:
        204:
          description: No content
        404:
          description: Not found
...
  /serie:
    get:
      tags: 
      - Serie
      summary: getAllSeries
      operationId: getAllSeries
      description: Devuelve todas las series
      produces: 
      - application/json
      responses:
        200:
          description: Devolviendo contenido
          schema:
            $ref: '#/definitions/Serie'
    post:
      tags:
      - Serie
      summary: addSerie
      parameters:
      - in: body
        name: addSerie
        description: Añadir serie
        schema:
          $ref: '#/definitions/Serie'
      operationId: addSerie
      description: Añadir una serie
      produces:
      - application/json
      responses:
        201:
          description: Created
    put:
      tags:
      - Serie
      summary: updateSerie
      parameters:
      - in: body
        name: updateSerie
        description: Actualizar serie
        schema:
          $ref: '#/definitions/Serie'
      operationId: updateSerie
      description: Actualizar serie
      produces:
      - application/json
      responses:
        204:
          description: No contend
        404:
          description: Not found
  /serie/{serieId}:
    get:
      tags:
      - Serie
      summary: getSerie
      parameters:
      - in: path
        name: serieId
        required: true
        description: Id de serie
        schema:
          type: string
      operationId: getSerie
      description: Devuelve una serie
      produces:
      - application/json
      responses:
        200:
          description: Devuelve la serie
          schema:
            $ref: '#/definitions/Serie'
    delete:
      tags:
      - Serie
      summary: deleteSerie
      parameters:
      - in: path
        name: serieId
        required: true
        description: Id de serie
        schema:
          type: string
      operationId: deleteSerie
      description: Elimina una serie
      produces:
      - application/json
      responses:
        204:
          description: No content
        404:
          description: Not found
  /pelicula:
    get:
      tags: 
      - Pelicula
      summary: getAllPeliculas
      operationId: getAllPeliculas
      description: Devuelve todas las peliculas
      produces: 
      - application/json
      responses:
        200:
          description: Devolviendo contenido
          schema:
            $ref: '#/definitions/Pelicula'
    post:
      tags:
      - Pelicula
      summary: addPelicula
      parameters:
      - in: body
        name: addPelicula
        description: Añadir pelicula
        schema:
          $ref: '#/definitions/Pelicula'
      operationId: addPelicula
      description: Añadir una pelicula
      produces:
      - application/json
      responses:
        201:
          description: Created
    put:
      tags:
      - Pelicula
      summary: updatePelicula
      parameters:
      - in: body
        name: updatePelicula
        description: Actualizar pelicula
        schema:
          $ref: '#/definitions/Pelicula'
      operationId: updatePelicula
      description: Actualizar pelicula
      produces:
      - application/json
      responses:
        204:
          description: No contend
        404:
          description: Not found
  /pelicula/{peliculaId}:
    get:
      tags:
      - Pelicula
      summary: getPelicula
      parameters:
      - in: path
        name: peliculaId
        required: true
        description: Id de pelicula
        schema:
          type: string
      operationId: getPelicula
      description: Devuelve una pelicula
      produces:
      - application/json
      responses:
        200:
          description: Devuelve la pelicula
          schema:
            $ref: '#/definitions/Pelicula'
    delete:
      tags:
      - Pelicula
      summary: deletePelicula
      parameters:
      - in: path
        name: peliculaId
        required: true
        description: Id de pelicula
        schema:
          type: string
      operationId: deletePelicula
      description: Elimina una pelicula
      produces:
      - application/json
      responses:
        204:
          description: No content
        404:
          description: Not found
  /tipo:
    post:
      tags:
      - Tipo
      summary: addTipo
      parameters:
      - in: body
        name: addTipo
        description: Añadir tipo
        schema:
          $ref: '#/definitions/Tipo'
      operationId: addTipo
      description: Añadir un tipo
      produces:
      - application/json
      responses:
        201:
          description: Created
  /tipo/{tipoId}:
    get:
      tags:
      - Tipo
      summary: getTipo
      parameters:
      - in: path
        name: tipoId
        required: true
        description: Id de tipo
        schema:
          type: string
      operationId: getTipo
      description: Devuelve un tipo
      produces:
      - application/json
      responses:
        200:
          description: Devuelve el tipo
          schema:
            $ref: '#/definitions/Tipo'
  /tipo/{tipoId}/{peliculaId}:
    post:
      tags:
      - Tipo
      summary: addPeliculaATipo
      parameters:
      - in: path
        name: tipoId
        required: true
        schema:
          type: string
        description: Id del tipo
      - in: path
        name: peliculaId
        required: true
        schema:
          type: string
        description: Id de la pelicula
      operationId: addPeliculaATipo
      description: Añade pelicula a tipo
      produces:
      - application/json
      responses:
        201:
          description: Created
        400:
          description: Bad request
        404:
          description: Not found
    delete:
      tags:
      - Tipo
      summary: deletePeliculaDeTipo
      parameters:
      - in: path
        name: tipoId
        required: true
        schema:
          type: string
        description: Id del tipo
      - in: path
        name: peliculaId
        required: true
        schema:
          type: string
      operationId: deletePeliculaDeTipo
      description: Elimina pelicula del tipo
      produces:
      - application/json
      responses:
        204:
          description: No contend
        404:
          description: Not found
  /tipo/{tipoId}/{serieId}:
    post:
      tags:
      - Tipo
      summary: addSerieATipo
      parameters:
      - in: path
        name: tipoId
        required: true
        schema:
          type: string
        description: Id del tipo
      - in: path
        name: serieId
        required: true
        schema:
          type: string
        description: Id de la serie
      operationId: addSerieATipo
      description: Añade serie a tipo
      produces:
      - application/json
      responses:
        201:
          description: Created
        400:
          description: Bad request
        404:
          description: Not found
    delete:
      tags:
      - Tipo
      summary: deleteSerieDeTipo
      parameters:
      - in: path
        name: tipoId
        required: true
        schema:
          type: string
        description: Id del tipo
      - in: path
        name: serieId
        required: true
        schema:
          type: string
      operationId: deleteSerieDeTipo
      description: Elimina serie del tipo
      produces:
      - application/json
      responses:
        204:
          description: No contend
        404:
          description: Not found
...
definitions:
...
  Song:
    required:
    - id
    - title
    properties:
      id:
        type: string
        example: so1
      title:
        type: string
        example: Show must go on
      artist:
        type: string
        example: Queen
      album:
        type: string
        example: Bohemian Rhapsody
      year:
        type: string
        example: 1970
  Tipo:
    required:
    - id
    properties:
      id:
        type: string
        example: t1
      serie:
        $ref: '#/definitions/Serie'
      pelicula:
        $ref: '#/definitions/Pelicula'
  Pelicula:
    required:
    - id
    - director
    properties:
      id:
        type: string
        example: pe1
      director:
        type: string
        example: Cristopher Nolan
      fechaEstreno:
        type: string
        format: date
        example: 15-11-2019
  Serie:
    required:
    - id
    - temporadas
    properties:
      id:
        type: string
        example: s1
      temporadas:
        type: integer
        example: 2
      ultFechaEmision:
        type: string
        format: date
        example: 20-10-2017
      estreno:
        type: string
        format: date
        example: 12-9-2013
      canal:
        type: string
        example: HBO
# Added by API Auto Mocking Plugin
# host: snmanager.appspot.com
# basePath: /api/
schemes:
 - https
# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
basePath: /alvagualh/MovieFinder/1.0.0